cts misspelling

+++ src/cts/include/cts/TritonCTS.h
@@ -113,7 +113,7 @@ private:
   void removeNonClockNets();
   void computeITermPosition(odb::dbITerm* term, int& x, int& y) const;
   void countSinksPostDbWrite(odb::dbNet* net, unsigned &sinks, unsigned & leafSinks,
-                             unsigned currWireLength, double &sinkWireLength,
+                             double currWireLength, double &sinkWireLength,
                              int& minDepth, int& maxDepth, int depth,
                              bool fullTree = false);
   std::pair<int, int> branchBufferCount(ClockInst* inst,
